+++ cedar-policy/CHANGELOG.md
@@ -29,6 +29,8 @@ Cedar Language Version: TBD
 - Added `to_cedar` functions for `PolicySet`, `Policy`, and `Template` that
   render the policy in the human-readable Cedar syntax. These functions can be used
   to convert JSON formatted policies into the human-readable syntax.
+- Added `Validator::schema()` to get a reference to the `Schema` even after it has been
+  consumed to construct a `Validator` (#1524)
 
 ## [4.3.3] - 2025-02-25